requirements.txt
requests==2.11.1
fuzzywuzzy==0.18.0
pymongo==3.10.1
wikipedia==1.4.0
python-telegram-bot
python-Levenshtein